Cheap/hacky fix for Linux crashes.
Pending complete and safe fix.
This is a weird one. Happens on random netmessages vtables and on random
indices, usually constant on a specific machine. I don't believe it is
the memory protection since it has happened on indices above 0. Maybe
we're writing on bogus memory (check if we're actually writing on the
vtable and not somewhere else).